requirements.txt
openai
pinecone-client
python-dotenv
streamlit